pyproject.toml
[project]
name = "truncated-quantile-critics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium>=1.2.3",
    "hydra-core>=1.3.2",
    "numpy>=2.4.4",
    "pytorch-lightning>=2.6.1",
    "torch>=2.11.0",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "gymnasium[box2d]>=1.2.3",
    "ruff>=0.15.9",
]
